+++ TimeStamper.sublime-settings
@@ -1,0 +1,4 @@
+{
+    "accepted_paths"        : ["*"],
+    "accepted_extensions"   : ["js","css"]
+}
