--- TimeStamper/TimeStamper.sublime-settings
@@ -1,4 +1,0 @@
-{
-    "accepted_paths"        : ["*"],
-    "accepted_extensions"   : ["js","css"]
-}
